# log logger's errors