@melekes 2017-11-06
```
BEFORE:
BenchmarkTMLoggerSimple-2                 200000              5830 ns/op             752 B/op       17 allocs/op
BenchmarkTMLoggerContextual-2             200000              6399 ns/op            1104 B/op       23 allocs/op
BenchmarkTMFmtLoggerSimple-2              300000              4953 ns/op             273 B/op        9 allocs/op
BenchmarkTMFmtLoggerContextual-2          300000              5384 ns/op             481 B/op       13 allocs/op

AFTER:
BenchmarkTMLoggerSimple-2                 200000              5859 ns/op             752 B/op       17 allocs/op
BenchmarkTMLoggerContextual-2             200000              6498 ns/op            1104 B/op       23 allocs/op
BenchmarkTMFmtLoggerSimple-2              300000              4935 ns/op             273 B/op        9 allocs/op
BenchmarkTMFmtLoggerContextual-2          200000              5668 ns/op             481 B/op       13 allocs/op
```